requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-whisperer==0.2.15
Django==4.2.5
filelock==3.12.4
idna==3.4
Jinja2==3.1.2
kombu==5.3.2
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
openai-whisper @ git+https://github.com/openai/whisper.git@0a60fcaa9b86748389a656aa013c416030287d47
prompt-toolkit==3.0.39
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
rabbitmq==0.2.0
rabbitmq-server==0.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
